+++ History.md
@@ -1,4 +1,9 @@
+0.1.1
+===================
+    * add method getAllCookies
+    * support tough-cookie 2.0.0
+
 0.1.0
 ===================
 
-  * Release
+    * Release